Add files via upload
In ST 307, I was given two parts for this project. For the first one, I was tasked with creating scatterplots with a MLR model. For the second part, I was tasked with a dataset as a marketer to analyze data from New York to understand where consumers make their purchases. I was able to find that there was a relationship between the independent and dependent variables.